[MEDI-58] refactor: refactoring notification API. 7/31.

+++ src/main/java/com/my_medi/domain/report/entity/Measurement.java
@@ -14,8 +14,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Measurement {
-    private Integer height; // cm
-    private Integer weight; // kg
+    private Double height; // cm
+    private Double weight; // kg
     private Double bmi;
     @Enumerated(EnumType.STRING)
     private BmiCategory bmiCategory;
